requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
boto3==1.35.29
botocore==1.35.29
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
docstring_parser==0.16
einops==0.8.0
encodec==0.1.1
fastapi==0.115.0
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
funcy==2.0
gradio==5.0.1
gradio_client==1.4.0
granian==1.4.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
httpx-ws==0.6.2
huggingface-hub==0.25.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
litellm==1.48.6
loguru==0.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.3
numpy==2.1.1
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
promptic==0.7.5
psutil==5.9.8
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
pypdf==4.1.0
python-dateutil==2.9.0.post0
python-dotenv>=1.0.0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests>=2.31.0
rich==13.8.1
rpds-py==0.20.0
ruff==0.6.8
s3transfer==0.10.2
safetensors==0.4.5
scipy==1.14.1
semantic-version==2.10.0
sentry-sdk==2.5.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
spaces==0.30.2
starlette==0.38.6
suno-bark @ git+https://github.com/suno-ai/bark.git@f4f32d4cd480dfec1c245d258174bc9bde3c2148
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.7.0
tokenizers==0.20.0
tomlkit==0.12.0
torch==2.4.1
torchaudio==2.4.1
tqdm==4.66.5
transformers==4.45.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3>=2.0.7
uvicorn==0.31.0
uvloop==0.18.0
websockets==12.0
wsproto==1.2.0
yarl==1.13.1
zipp==3.20.2
backoff>=2.2.1
